Move CLI to rune package
