+++ crates/rune-cli/src/main.rs
@@ -29,663 +29,19 @@
 //! [Rune Language]: https://rune-rs.github.io
 //! [rune]: https://github.com/rune-rs/rune
 
-mod benches;
-mod check;
-mod doc;
-mod loader;
-mod run;
-mod tests;
-mod visitor;
-
-use anyhow::{bail, Result};
-use clap::{Parser, Subcommand};
-use rune::compile::{ItemBuf, ParseOptionError};
-use rune::termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
-use rune::workspace::WorkspaceFilter;
-use rune::{Context, ContextError, Options};
-use rune_modules::capture_io::CaptureIo;
-use std::error::Error;
-use std::io::{self, Write};
-use std::path::{Path, PathBuf};
-use tracing_subscriber::filter::EnvFilter;
-
 pub const VERSION: &str = include_str!(concat!(env!("OUT_DIR"), "/version.txt"));
 
-struct EntryPoint {
-    item: ItemBuf,
-    paths: Vec<PathBuf>,
-}
+fn main() {
+    rune::cli::Entry::new()
+        .about(format_args!("The Rune Language Interpreter {VERSION}"))
+        .context(&mut |opts| {
+            let mut c = rune_modules::with_config(opts.capture.is_none())?;
 
-struct Io<'a> {
-    stdout: &'a mut StandardStream,
-    stderr: &'a mut StandardStream,
-}
-
-#[derive(Subcommand, Debug, Clone)]
-enum Command {
-    /// Run checks but do not execute
-    Check(check::Flags),
-    /// Build documentation.
-    Doc(doc::Flags),
-    /// Run all tests but do not execute
-    Test(tests::Flags),
-    /// Run the given program as a benchmark
-    Bench(benches::Flags),
-    /// Run the designated script
-    Run(run::Flags),
-}
-
-impl Command {
-    fn propagate_related_flags(&mut self, c: &mut Config) {
-        match self {
-            Command::Check(..) => {}
-            Command::Doc(..) => {}
-            Command::Test(..) => {
-                c.test = true;
+            if opts.experimental {
+                c.install(rune_modules::experiments::module(opts.capture.is_none())?)?;
             }
-            Command::Bench(..) => {
-                c.test = true;
-            }
-            Command::Run(args) => {
-                args.propagate_related_flags();
-            }
-        }
-    }
 
-    fn describe(&self) -> &'static str {
-        match self {
-            Command::Check(..) => "Checking",
-            Command::Doc(..) => "Building documentation",
-            Command::Test(..) => "Testing",
-            Command::Bench(..) => "Benchmarking",
-            Command::Run(..) => "Running",
-        }
-    }
-
-    fn shared(&self) -> &SharedFlags {
-        match self {
-            Command::Check(args) => &args.shared,
-            Command::Doc(args) => &args.shared,
-            Command::Test(args) => &args.shared,
-            Command::Bench(args) => &args.shared,
-            Command::Run(args) => &args.shared,
-        }
-    }
-
-    fn bins_test(&self) -> Option<WorkspaceFilter<'_>> {
-        if !matches!(
-            self,
-            Command::Run(..) | Command::Check(..) | Command::Doc(..)
-        ) {
-            return None;
-        }
-
-        let shared = self.shared();
-
-        Some(if let Some(name) = &shared.bin {
-            WorkspaceFilter::Name(name)
-        } else {
-            WorkspaceFilter::All
+            Ok(c)
         })
-    }
-
-    fn tests_test(&self) -> Option<WorkspaceFilter<'_>> {
-        if !matches!(
-            self,
-            Command::Test(..) | Command::Check(..) | Command::Doc(..)
-        ) {
-            return None;
-        }
-
-        let shared = self.shared();
-
-        Some(if let Some(name) = &shared.test {
-            WorkspaceFilter::Name(name)
-        } else {
-            WorkspaceFilter::All
-        })
-    }
-
-    fn examples_test(&self) -> Option<WorkspaceFilter<'_>> {
-        if !matches!(
-            self,
-            Command::Run(..) | Command::Check(..) | Command::Doc(..)
-        ) {
-            return None;
-        }
-
-        let shared = self.shared();
-
-        Some(if let Some(name) = &shared.example {
-            WorkspaceFilter::Name(name)
-        } else {
-            WorkspaceFilter::All
-        })
-    }
-
-    fn benches_test(&self) -> Option<WorkspaceFilter<'_>> {
-        if !matches!(
-            self,
-            Command::Bench(..) | Command::Check(..) | Command::Doc(..)
-        ) {
-            return None;
-        }
-
-        let shared = self.shared();
-
-        Some(if let Some(name) = &shared.bench {
-            WorkspaceFilter::Name(name)
-        } else {
-            WorkspaceFilter::All
-        })
-    }
-}
-
-#[derive(Parser, Debug, Clone)]
-struct SharedFlags {
-    /// Enable experimental features.
-    ///
-    /// This makes the `std::experimental` module available to scripts.
-    #[arg(long)]
-    experimental: bool,
-
-    /// Recursively load all files in the given directory.
-    #[arg(long)]
-    recursive: bool,
-
-    /// Display warnings.
-    #[arg(long)]
-    warnings: bool,
-
-    /// Set the given compiler option (see `--help` for available options).
-    ///
-    /// memoize-instance-fn[=<true/false>] - Inline the lookup of an instance function where appropriate.
-    ///
-    /// link-checks[=<true/false>] - Perform linker checks which makes sure that called functions exist.
-    ///
-    /// debug-info[=<true/false>] - Enable or disable debug info.
-    ///
-    /// macros[=<true/false>] - Enable or disable macros (experimental).
-    ///
-    /// bytecode[=<true/false>] - Enable or disable bytecode caching (experimental).
-    #[arg(name = "option", short = 'O', number_of_values = 1)]
-    compiler_options: Vec<String>,
-
-    /// Run with the following binary from a loaded manifest. This requires a
-    /// `Rune.toml` manifest.
-    #[arg(long = "bin")]
-    bin: Option<String>,
-
-    /// Run with the following test from a loaded manifest. This requires a
-    /// `Rune.toml` manifest.
-    #[arg(long = "test")]
-    test: Option<String>,
-
-    /// Run with the following example from a loaded manifest. This requires a
-    /// `Rune.toml` manifest.
-    #[arg(long = "example")]
-    example: Option<String>,
-
-    /// Run with the following benchmark by name from a loaded manifest. This
-    /// requires a `Rune.toml` manifest.
-    #[arg(long = "bench")]
-    bench: Option<String>,
-
-    /// All paths to include in the command. By default, the tool searches the
-    /// current directory and some known files for candidates.
-    paths: Vec<PathBuf>,
-}
-
-struct Package {
-    /// The name of the package the path belongs to.
-    name: String,
-}
-
-enum Entry {
-    /// A plain path entry.
-    Path(PathBuf),
-    /// A path from a specific package.
-    PackagePath(Package, PathBuf),
-}
-
-#[derive(Default)]
-struct Config {
-    /// Whether or not the test module should be included.
-    test: bool,
-    /// Whether or not to use verbose output.
-    verbose: bool,
-    /// Manifest root directory.
-    manifest_root: Option<PathBuf>,
-    /// The explicit paths to load.
-    entries: Vec<Entry>,
-}
-
-impl SharedFlags {
-    /// Construct a rune context according to the specified argument.
-    fn context(&self, c: &Config) -> Result<Context, ContextError> {
-        let mut context = rune_modules::default_context()?;
-
-        if self.experimental {
-            context.install(rune_modules::experiments::module(true)?)?;
-        }
-
-        if c.test {
-            context.install(benches::test_module()?)?;
-        }
-
-        Ok(context)
-    }
-
-    /// Setup a context that captures output.
-    fn context_with_capture(&self, c: &Config, io: &CaptureIo) -> Result<Context, ContextError> {
-        let mut context = rune_modules::with_config(false)?;
-
-        context.install(rune_modules::capture_io::module(io)?)?;
-
-        if self.experimental {
-            context.install(rune_modules::experiments::module(true)?)?;
-        }
-
-        if c.test {
-            context.install(benches::test_module()?)?;
-        }
-
-        Ok(context)
-    }
-}
-
-#[derive(Parser, Debug, Clone)]
-#[command(name = "rune", about = "The Rune Language Interpreter", version = VERSION)]
-struct Args {
-    /// Control if output is colored or not.
-    ///
-    /// Valid options are:
-    /// * `auto` - try to detect automatically.
-    /// * `ansi` - unconditionally emit ansi control codes.
-    /// * `always` - always enabled.
-    ///
-    /// Anything else will disable coloring.
-    #[arg(short = 'C', long, default_value = "auto")]
-    color: String,
-
-    /// The command to execute
-    #[command(subcommand)]
-    cmd: Command,
-}
-
-impl Args {
-    /// Construct compiler options from cli arguments.
-    fn options(&self) -> Result<Options, ParseOptionError> {
-        let mut options = Options::default();
-
-        // Command-specific override defaults.
-        match &self.cmd {
-            Command::Test(_) | Command::Check(_) => {
-                options.debug_info(true);
-                options.test(true);
-                options.bytecode(false);
-            }
-            Command::Bench(_) | Command::Doc(..) | Command::Run(_) => (),
-        }
-
-        for option in &self.cmd.shared().compiler_options {
-            options.parse_option(option)?;
-        }
-
-        Ok(options)
-    }
-}
-
-const SPECIAL_FILES: &[&str] = &[
-    "main.rn",
-    "lib.rn",
-    "src/main.rn",
-    "src/lib.rn",
-    "script/main.rn",
-    "script/lib.rn",
-];
-
-// Our own private ExitCode since std::process::ExitCode is nightly only.
-// Note that these numbers are actually meaningful on Windows, but we don't
-// care.
-#[repr(i32)]
-enum ExitCode {
-    Success = 0,
-    Failure = 1,
-    VmError = 2,
-}
-
-#[tokio::main]
-async fn main() {
-    match try_main().await {
-        Ok(exit_code) => {
-            std::process::exit(exit_code as i32);
-        }
-        Err(error) => {
-            let o = std::io::stderr();
-            // ignore error because stdout / stderr might've been closed.
-            let _ = format_errors(o.lock(), &error);
-            std::process::exit(-1);
-        }
-    }
-}
-
-/// Format the given error.
-fn format_errors<O>(mut o: O, error: &dyn Error) -> io::Result<()>
-where
-    O: io::Write,
-{
-    writeln!(o, "Error: {}", error)?;
-    let mut source = error.source();
-
-    while let Some(error) = source.take() {
-        writeln!(o, "Caused by: {}", error)?;
-        source = error.source();
-    }
-
-    Ok(())
-}
-
-async fn try_main() -> Result<ExitCode, io::Error> {
-    let args = match Args::try_parse() {
-        Ok(args) => args,
-        Err(e) => {
-            let code = if e.use_stderr() {
-                let mut o = std::io::stderr();
-                writeln!(o, "{}", e)?;
-                ExitCode::Failure
-            } else {
-                let mut o = std::io::stdout();
-                writeln!(o, "{}", e)?;
-                ExitCode::Success
-            };
-
-            return Ok(code);
-        }
-    };
-
-    let choice = match args.color.as_str() {
-        "always" => ColorChoice::Always,
-        "ansi" => ColorChoice::AlwaysAnsi,
-        "auto" => {
-            if atty::is(atty::Stream::Stdout) {
-                ColorChoice::Auto
-            } else {
-                ColorChoice::Never
-            }
-        }
-        "never" => ColorChoice::Never,
-        _ => ColorChoice::Auto,
-    };
-
-    let mut stdout = StandardStream::stdout(choice);
-    let mut stderr = StandardStream::stderr(choice);
-
-    let mut io = Io {
-        stdout: &mut stdout,
-        stderr: &mut stderr,
-    };
-
-    tracing_subscriber::fmt()
-        .with_env_filter(EnvFilter::from_default_env())
-        .init();
-
-    match main_with_out(&mut io, args).await {
-        Ok(code) => Ok(code),
-        Err(error) => {
-            let mut o = io.stdout.lock();
-            o.set_color(ColorSpec::new().set_fg(Some(Color::Red)))?;
-            let result = format_errors(&mut o, error.as_ref());
-            o.set_color(&ColorSpec::new())?;
-            result?;
-            Ok(ExitCode::Failure)
-        }
-    }
-}
-
-fn find_manifest() -> Result<(PathBuf, PathBuf)> {
-    let mut path = PathBuf::new();
-
-    loop {
-        let manifest_path = path.join(rune::workspace::MANIFEST_FILE);
-
-        if manifest_path.is_file() {
-            return Ok((path, manifest_path));
-        }
-
-        path.push("..");
-
-        if !path.is_dir() {
-            bail!(
-                "coult not find {} in this or parent directories",
-                rune::workspace::MANIFEST_FILE
-            )
-        }
-    }
-}
-
-fn populate_config(io: &mut Io<'_>, c: &mut Config, args: &Args) -> Result<()> {
-    c.entries.extend(
-        args.cmd
-            .shared()
-            .paths
-            .iter()
-            .map(|p| Entry::Path(p.as_path().into())),
-    );
-
-    if !c.entries.is_empty() {
-        return Ok(());
-    }
-
-    for file in SPECIAL_FILES {
-        let path = Path::new(file);
-
-        if path.is_file() {
-            c.entries.push(Entry::Path(path.into()));
-            return Ok(());
-        }
-    }
-
-    let (manifest_root, manifest_path) = find_manifest()?;
-
-    // When building or running a workspace we need to be more verbose so that
-    // users understand what exactly happens.
-    c.verbose = true;
-    c.manifest_root = Some(manifest_root);
-
-    let mut sources = rune::Sources::new();
-    sources.insert(rune::Source::from_path(&manifest_path)?);
-
-    let mut diagnostics = rune::workspace::Diagnostics::new();
-
-    let result = rune::workspace::prepare(&mut sources)
-        .with_diagnostics(&mut diagnostics)
-        .build();
-
-    diagnostics.emit(io.stdout, &sources)?;
-
-    let manifest = result?;
-
-    if let Some(bin) = args.cmd.bins_test() {
-        for found in manifest.find_bins(bin)? {
-            let package = Package {
-                name: found.package.name.clone(),
-            };
-            c.entries.push(Entry::PackagePath(package, found.path));
-        }
-    }
-
-    if let Some(test) = args.cmd.tests_test() {
-        for found in manifest.find_tests(test)? {
-            let package = Package {
-                name: found.package.name.clone(),
-            };
-            c.entries.push(Entry::PackagePath(package, found.path));
-        }
-    }
-
-    if let Some(example) = args.cmd.examples_test() {
-        for found in manifest.find_examples(example)? {
-            let package = Package {
-                name: found.package.name.clone(),
-            };
-            c.entries.push(Entry::PackagePath(package, found.path));
-        }
-    }
-
-    if let Some(bench) = args.cmd.benches_test() {
-        for found in manifest.find_benches(bench)? {
-            let package = Package {
-                name: found.package.name.clone(),
-            };
-            c.entries.push(Entry::PackagePath(package, found.path));
-        }
-    }
-
-    Ok(())
-}
-
-async fn main_with_out(io: &mut Io<'_>, mut args: Args) -> Result<ExitCode> {
-    let mut c = Config::default();
-    args.cmd.propagate_related_flags(&mut c);
-    populate_config(io, &mut c, &args)?;
-
-    let entries = std::mem::take(&mut c.entries);
-    let options = args.options()?;
-
-    let what = args.cmd.describe();
-    let verbose = c.verbose;
-    let recursive = args.cmd.shared().recursive;
-
-    let mut entrys = Vec::new();
-
-    for entry in &entries {
-        let (item, path) = match entry {
-            Entry::Path(path) => (ItemBuf::new(), path),
-            Entry::PackagePath(p, path) => {
-                if verbose {
-                    let mut o = io.stderr.lock();
-                    o.set_color(ColorSpec::new().set_fg(Some(Color::Green)).set_bold(true))?;
-                    let result = write!(o, "{:>12}", what);
-                    o.set_color(&ColorSpec::new())?;
-                    result?;
-                    writeln!(o, " `{}` (from {})", path.display(), p.name)?;
-                }
-
-                (ItemBuf::with_crate(&p.name), path)
-            }
-        };
-
-        let mut paths = Vec::new();
-
-        for path in loader::recurse_paths(recursive, path.clone()) {
-            paths.push(path?);
-        }
-
-        entrys.push(EntryPoint { item, paths });
-    }
-
-    match run_path(io, &c, &args, &options, entrys).await? {
-        ExitCode::Success => (),
-        other => {
-            return Ok(other);
-        }
-    }
-
-    Ok(ExitCode::Success)
-}
-
-/// Run a single path.
-async fn run_path<I>(
-    io: &mut Io<'_>,
-    c: &Config,
-    args: &Args,
-    options: &Options,
-    entrys: I,
-) -> Result<ExitCode>
-where
-    I: IntoIterator<Item = EntryPoint>,
-{
-    match &args.cmd {
-        Command::Check(flags) => {
-            for e in entrys {
-                for path in &e.paths {
-                    match check::run(io, c, flags, options, path)? {
-                        ExitCode::Success => (),
-                        other => return Ok(other),
-                    }
-                }
-            }
-        }
-        Command::Doc(flags) => return doc::run(io, c, flags, options, entrys),
-        Command::Test(flags) => {
-            for e in entrys {
-                for path in &e.paths {
-                    let capture_io = rune_modules::capture_io::CaptureIo::new();
-                    let context = flags.shared.context_with_capture(c, &capture_io)?;
-
-                    let load =
-                        loader::load(io, &context, args, options, path, visitor::Attribute::Test)?;
-
-                    match tests::run(
-                        io,
-                        flags,
-                        &context,
-                        Some(&capture_io),
-                        load.unit,
-                        &load.sources,
-                        &load.functions,
-                    )
-                    .await?
-                    {
-                        ExitCode::Success => (),
-                        other => return Ok(other),
-                    }
-                }
-            }
-        }
-        Command::Bench(flags) => {
-            for e in entrys {
-                for path in &e.paths {
-                    let capture_io = rune_modules::capture_io::CaptureIo::new();
-                    let context = flags.shared.context_with_capture(c, &capture_io)?;
-
-                    let load =
-                        loader::load(io, &context, args, options, path, visitor::Attribute::Bench)?;
-
-                    match benches::run(
-                        io,
-                        flags,
-                        &context,
-                        Some(&capture_io),
-                        load.unit,
-                        &load.sources,
-                        &load.functions,
-                    )
-                    .await?
-                    {
-                        ExitCode::Success => (),
-                        other => return Ok(other),
-                    }
-                }
-            }
-        }
-        Command::Run(flags) => {
-            let context = flags.shared.context(c)?;
-
-            for e in entrys {
-                for path in &e.paths {
-                    let load =
-                        loader::load(io, &context, args, options, path, visitor::Attribute::None)?;
-
-                    match run::run(io, c, flags, &context, load.unit, &load.sources).await? {
-                        ExitCode::Success => (),
-                        other => return Ok(other),
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(ExitCode::Success)
+        .run();
 }
