+++ crates/rune-wasm/Cargo.toml
@@ -23,7 +23,7 @@ gloo-utils = "0.1.6"
 
 rune = { version = "0.12.3", path = "../rune" }
 rune-macros = { version = "=0.12.3", path = "../rune-macros" }
-rune-modules = { version = "0.12.3", path = "../rune-modules", features = ["core", "test", "json", "toml", "rand", "experiments", "macros", "capture-io"] }
+rune-modules = { version = "0.12.3", path = "../rune-modules", features = ["core", "test", "json", "toml", "rand", "experiments"] }
 
 [dependencies.web-sys]
 version = "0.3.61"
